+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.25] - 2025-05-15
+
+### Added
+
+- Commissions API endpoints:
+  - List commission groups
+  - Create a commission group
+  - Update a commission group
+  - Delete a commission group
+
 ## [1.0.0-beta.24] - 2025-04-23
 
 ### Fixed